🩹 Fix: wish 코드 수정
